+++ BattlegroundNumbers/Locale/enUS.lua
@@ -12,8 +12,6 @@ L["EnemyNameplates_Color_Enabled"] = "Enable Custom Color"
 L["EnemyNameplates_Color_Enabled_Desc"] = "Enables custom enemy name color"
 L["EnemyNameplates_Color"] = "Enemy Nameplate Names Color"
 L["EnemyNameplates_Color_Desc"] = "The color for enemy nameplate names when playing a battleground"
-L["AlliesNameplates_Color"] = "Allies Nameplate Names Color"
-L["AlliesNameplates_Color_Desc"] = "The color for allies nameplate names when playing a battleground"
 L["EnemyNameplates_Format"] = "Name Format"
 L["EnemyNameplates_Format_Desc"] = [=[Format for nameplate name. ie.: CLASS SPEC - NUM
 
@@ -23,3 +21,18 @@ CLASS - player class
 SPEC - player spec
 NUM - number (classwide)
 ]=]
+L["AllyNameplates"] = "Ally Nameplates"
+L["AllyNameplates_Color_Always"] = "Enable Custom Color Always"
+L["AllyNameplates_Color_Always_Desc"] = "Enables custom ally name color always. Including world."
+L["AllyNameplates_Color"] = "Ally Nameplate Names Color"
+L["AllyNameplates_Color_Desc"] = "The color for ally nameplate names"
+L["AllyNameplates_Color_Enabled"] = "Enable Custom Color"
+L["AllyNameplates_Color_Enabled_Desc"] = "Enables custom ally name color"
+L["AllyNameplates_Color_Arena"] = "Arenas"
+L["AllyNameplates_Color_BG"] = "BattleGrounds"
+L["AllyNameplates_Hide_HealthBar_Always"] = "Hide Health Bar Always"
+L["AllyNameplates_Hide_HealthBar_Always_Desc"] = "Hide Ally nameplate health bars always. Including world."
+L["AllyNameplates_Hide_HealthBar_Enabled"] = "Hide Health Bar"
+L["AllyNameplates_Hide_HealthBar_Enabled_Desc"] = "Hide ally nameplate health bar"
+L["AllyNameplates_Hide_HealthBar_Arena"] = "Arenas"
+L["AllyNameplates_Hide_HealthBar_BG"] = "BattleGrounds"